+++ packages/imperative/src/plugins/utilities/npm-interface/install.ts
@@ -19,6 +19,9 @@ import { ImperativeError } from "../../../../../error";
 import { IPluginJsonObject } from "../../doc/IPluginJsonObject";
 import { getPackageInfo, installPackages } from "../NpmFunctions";
 import { ConfigSchema } from "../../../../../config/src/ConfigSchema";
+import { PluginManagementFacility } from "../../PluginManagementFacility";
+import { ConfigurationLoader } from "../../../ConfigurationLoader";
+import { UpdateImpConfig } from "../../../UpdateImpConfig";
 
 
 /**
@@ -128,6 +131,14 @@ export async function install(packageLocation: string, registry: string, install
         iConsole.debug(`Checking for global team configuration files to update.`);
         if (PMFConstants.instance.PLUGIN_USING_CONFIG &&
             PMFConstants.instance.PLUGIN_CONFIG.layers.filter((layer) => layer.global && layer.exists).length > 0) {
+            // Update the Imperative Configuration to add the profiles introduced by the recently installed plugin
+            // This might be needed outside of PLUGIN_USING_CONFIG scenarios, but we haven't had issues with other APIs before
+            const pmfInstance = PluginManagementFacility.instance;
+            // TODO: Hacky way to pass the plugin name to the requirePluginModuleCallback function
+            (pmfInstance as any).pluginNmForUseInCallback = packageInfo.name;
+            const pluginConfig = ConfigurationLoader.load(null, packageInfo, pmfInstance.requirePluginModuleCallback.bind(pmfInstance));
+            UpdateImpConfig.addProfiles(pluginConfig.profiles);
+
             ConfigSchema.updateSchema({ layer: "global" });
         }